# Swagger/swagger.yaml
swagger: '2.0'
info:
  description: ''
  version: '2'
  title: MVC
host: 'localhost:4444'
tags:

  - name: Bank

schemes:
  - http
paths:
  /api/deposit:
    post:
      tags:
        - Bank
      summary: Deposit
      description: Deposit your money
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: account_number
          in: formData
          description: เลขบัญชี
          required: true
          type: string
          default: "2562342110003"
        - name: pin
          in: formData
          description: รหัส4หลัก
          required: true
          type: string
          default: "8143"
        - name: depo
          in: formData
          description: จำนวนเงินที่ต้องการฝาก
          required: true
          type: string
          default: "10000"
      responses:
        '201':
          description: update success
          schema:
            $ref: '#/definitions/createsuccess201'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/badRequest400'
        '401':
          description: unAuthenticate
          schema:
            $ref: '#/definitions/unAuthenticate401'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal500'
  /api/withdraw:
    post:
      tags:
        - Bank
      summary: Withdraw
      description: Withdraw your money
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: account_number
          in: formData
          description: เลขบัญชี
          required: true
          type: string
          default: "2562342110003"
        - name: pin
          in: formData
          description: รหัส4หลัก
          required: true
          type: string
          default: "8143"
        - name: withd
          in: formData
          description: จำนวนเงินที่ต้องการถอน
          required: true
          type: string
          default: "10000"
      responses:
        '201':
          description: update success
          schema:
            $ref: '#/definitions/createsuccess201'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/badRequest400'
        '401':
          description: unAuthenticate
          schema:
            $ref: '#/definitions/unAuthenticate401'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal500'
  /api/tranfer:
    post:
      tags:
        - Bank
      summary: Tranfer
      description: Tranfer your money
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: ac1
          in: formData
          description: เลขบัญชีผู้โอน
          required: true
          type: string
          default: "2562342110003"
        - name: pi1
          in: formData
          description: รหัส4หลักผู้โอน
          required: true
          type: string
          default: "8143"
        - name: ac2
          in: formData
          description: เลขบัญชีผู้รับ
          required: true
          type: string
          default: "2562342110002"
        - name: tran
          in: formData
          description: จำนวนเงินที่ต้องการโอน
          required: true
          type: string
          default: "10000"
      responses:
        '201':
          description: update success
          schema:
            $ref: '#/definitions/createsuccess201'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/badRequest400'
        '401':
          description: unAuthenticate
          schema:
            $ref: '#/definitions/unAuthenticate401'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal500'
  /api/getid:
    post:
      tags:
        - Bank
      summary: Check balance
      description: Check your money balance
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: account_number
          in: formData
          description: เลขบัญชี
          required: true
          type: string
          default: "2562342110003"
        - name: pin
          in: formData
          description: รหัส4หลัก
          required: true
          type: string
          default: "8143"
      responses:
        '201':
          description: update success
          schema:
            $ref: '#/definitions/createsuccess201'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/badRequest400'
        '401':
          description: unAuthenticate
          schema:
            $ref: '#/definitions/unAuthenticate401'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal500'
definitions:
  createsuccess201:
    type: object
    properties:
      statusCode:
        type: number
        example: 201
      message:
        type: string
        example: update success
  badRequest400:
    type: object
    properties:
      statusCode:
        type: number
        example: 400
      message:
        type: string
        example: Bad Request
  unAuthenticate401:
    type: object
    properties:
      statusCode:
        type: number
        example: 401
      message:
        type: string
        example: Unauthorized
  internal500:
    type: object
    properties:
      statusCode:
        type: number
        example: 500
      message:
        type: string
        example: Internal Server Error